docs: Update README and clean up console logs
