+++ src/components/detail/ContentWrapper.tsx
@@ -1,6 +1,11 @@
 import { Box } from "@mui/material";
-import { WrapperPropsType } from "../../shared/types";
 
-export const ContentWrapper = ({ children, mt, mb }: WrapperPropsType) => {
+interface ContentWrapperProps {
+  children: React.ReactNode;
+  mt?: string;
+  mb?: string;
+}
+
+export const ContentWrapper = ({ children, mt, mb }: ContentWrapperProps) => {
   return <Box sx={{ display: "flex", justifyContent: "center", mb: mb ?? 2, mt: mt ?? 0 }}>{children}</Box>;
 };
