+++ src/components/home/GridItem.tsx
@@ -1,7 +1,14 @@
 import { Box } from "@mui/material";
-import { WrapperPropsType } from "../../shared/types";
+import { Pokemon } from "pokenode-ts";
+import { useNavigate } from "react-router-dom";
+import { TypeWrapper } from "../common/TypeWrapper";
 
-export const GridItem = ({ children, onClick }: WrapperPropsType) => {
+interface GridItemProps {
+  pokemon: Pokemon;
+}
+
+export const GridItem = ({ pokemon }: GridItemProps) => {
+  const navigate = useNavigate();
   return (
     <Box
       sx={{
@@ -12,9 +19,26 @@ export const GridItem = ({ children, onClick }: WrapperPropsType) => {
         flexDirection: "column",
         cursor: "pointer",
       }}
-      onClick={onClick}
+      onClick={() => {
+        navigate(`/${pokemon.id}`);
+      }}
     >
-      {children}
+      <Box
+        className="name"
+        sx={{
+          fontSize: "12px",
+          fontWeight: "bold",
+          color: "lightslategray",
+        }}
+      >
+        No. {String(pokemon.id).padStart(4, "0")}
+      </Box>
+      <Box className="name" sx={{ fontSize: "20px", fontWeight: "bold" }}>
+        {pokemon.name.charAt(0).toUpperCase() + pokemon.name.slice(1)}
+      </Box>
+      <img src={pokemon.sprites.front_default ?? undefined} alt={pokemon.name} />
+
+      <TypeWrapper pokemon={pokemon} />
     </Box>
   );
 };
