+++ src/components/icons/HomeIcon.tsx
@@ -1,4 +1,3 @@
-import React from "react";
 import SvgIcon, { SvgIconProps } from "@mui/material/SvgIcon";
 
 const HomeIcon = (props: SvgIconProps) => {
