+++ src/components/layout/Footer.tsx
@@ -1,4 +1,3 @@
-import React from "react";
 import Box from "@mui/material/Box";
 
 const Footer = () => {
