+++ src/components/layout/Layout.tsx
@@ -1,8 +1,7 @@
-import React from "react";
 import Header from "./Header";
 import Footer from "./Footer";
-import { Outlet } from "react-router-dom";
 import Box from "@mui/material/Box";
+import { Outlet } from "react-router-dom";
 
 const Layout = () => {
   return (
