+++ src/shared/Router.tsx
@@ -1,15 +1,15 @@
-import React from "react";
-import { Route, Routes } from "react-router-dom";
 import Home from "../pages/Home";
 import Detail from "../pages/Detail";
 import Layout from "../components/layout/Layout";
+import { Navigate, Route, Routes } from "react-router-dom";
 
 const Router = () => {
   return (
     <Routes>
       <Route element={<Layout />}>
-        <Route path="/" element={<Home />}></Route>
-        <Route path="/:id" element={<Detail />}></Route>
+        <Route index element={<Home />} />
+        <Route path="/:id" element={<Detail />} />
+        <Route path="*" element={<Navigate replace to="/" />} />
       </Route>
     </Routes>
   );
